Navbar links added
