+++ src/locales/ru/translation.json
@@ -1,7 +1,7 @@
 {
   "Search": "Поиск",
   "Sign In": "Войти",
-  "Sign Up": "Регистр",
+  "Sign Up": "Регистраться",
   "Store Location: Lincoln- 344, Illinois, Chicago, USA": "Местоположение магазина: Линкольн-344, Иллинойс, Чикаго, США.",
   "Home": "Дом",
   "Shopping cart": "Корзина"
